refactor: move the health check component to canary directory
