+++ src/components/Topology/Sidebar/ComponentChecks.tsx
@@ -5,8 +5,8 @@ import { AiFillHeart } from "react-icons/ai";
 import { getComponentChecks } from "../../../api/services/topology";
 import PillBadge from "../../../ui/Badge/PillBadge";
 import CollapsiblePanel from "../../../ui/CollapsiblePanel/CollapsiblePanel";
+import { CheckLink } from "../../Canary/HealthChecks/CheckLink";
 import EmptyState from "../../EmptyState";
-import { CheckLink } from "../../HealthChecks/CheckLink";
 import TextSkeletonLoader from "../../SkeletonLoader/TextSkeletonLoader";
 import { refreshButtonClickedTrigger } from "../../SlidingSideBar";
 import Title from "../../Title/title";
