+++ src/components/Topology/TopologyCard/index.tsx
@@ -6,9 +6,9 @@ import { getTopology } from "../../../api/services/topology";
 import { Topology } from "../../../api/types/topology";
 import { Size } from "../../../types";
 import AgentName from "../../Agents/AgentName";
+import { HealthChecksSummary } from "../../Canary/HealthChecksSummary";
+import { HealthSummary } from "../../Canary/HealthSummary";
 import { CustomScroll } from "../../CustomScroll";
-import { HealthChecksSummary } from "../../HealthChecksSummary";
-import { HealthSummary } from "../../HealthSummary";
 import { Icon } from "../../Icon";
 import IncidentCardSummary from "../../IncidentCardSummary";
 import TopologyCardSkeletonLoader from "../../SkeletonLoader/TopologyCardSkeletonLoader";
